Added IFileProducerListener Support to IFileOutputTool
The IFileOutputTool and IFileOutputJobBuilder now allow for setting an
IFileProducerListener instance. The IFileProducerListener then can be
informed about the created files at the end of an output job's
execution.

This also led to the creation of the enum EArchiveType which hosts the
zip archive file type. We thus have the basic support to notify about
the created zip archive out of the box.

In the future, the Boolean parameters which we can use to specify
whether output or input jobs expect zip archives will be replaced with
an EArchiveType instance, thus allowing for more archive types.

Also, we can now extend the existing FileOutputTools to notify listeners
about their created files. This will make it easier to integrate them
into other APIs.