Harmonized Document, Chart, and Graphic API to Facilitate Evaluator
The Document, Chart, and Graphics API have been harmonized. Graphic
output drivers for different formats (such as JPEG, PNG, EPS, PDF,
etc.) can be implemented independently and hidden behind interfaces
and the Graphic2D subclass Graphic. Chart drawing code can be
implemented independently and hidden behind interfaces. Document
generating code can be implemented independently and hidden behind
interfaces. Yet, these three independent APIs are now seamlessly
integrated in the Document API. The key is that all three APIs are
based on the Tool API: A tool is a thing that can perform a certain
job. A tool can generate a builder object which is used to configure
the job and to set stuff such as a logger for log output. The builder
then can generate the object representing the job (e.g., a document
output interface, a graphic output object, a chart definition object).
The Document API now can be configured to use a certain graphics
output driver and a certain chart generation tool. The document API can
seamlessly lead over to the Graphics API (if you draw a graphic) or to
the Chart API (if you include a chart). All APIs can share the same
palette of colors, fonts, and strokes to generate an overall uniform
outlook through the whole document.

So far, the Document API has been implemented for XHTML 1.0, a LaTeX
implementation will follow. The Graphics API supports PNG, JPEG, GIF,
and BMP (through Java's imageio API) and SVG, SVGZ, EPS, and PDF
(through wrapping around FreeHEP). The Chart API is so far implemented
once, in form of a wrapper for JFreeChart. The goal is that you can
later write the a document either as XHTML or LaTeX document without
any change to your code, that you can use different graphics formats
or chart implementations. All of this won't matter to the code which
writes the text or paints the figures.

The harmonizing of the three APIs became necessary as I am now working
on the experiment evaluation code. The aim here is to provide you a
single point where the whole system can be configured with a single
configuration file. The system will then use the document generation
strategy configured to it and automatically load the required classes.
At least for the document output code, this is now possible. For the
data input code, it still has to be done.