Error Logging Code now Centralized in ErrorUtils
The class ErrorUtils now holds a method logError. This method receives
a Throwable and a Logger as parameter, as well as a message String and
a boolean parameter "forceLog". If the Logger is not null and can log
messages of level SEVERE, we may log the error to it, if the error has
not yet been logged to that Logger or "forceLog" is true. Whenever we
log an error to a given Logger, we add a sentinel Exception to its list
of suppressed errors. This way we can check to which Loggers an error
has been logged for arbitrary types of exceptions. This way we can
elegantly avoid logging the same error time and again. Because of this,
we can now add calls to logError more liberally to the code without the
fear of congesting the logs.