More Detailed Error Messages in Structured I/O Tools
I try to provide more detailed error messages in the structured I/O
tools. This is achieved by adding a field "m_current" to the _IOJob
class which is set to the current file/path/uri/url/resource in the
corresponding handler routines and reset in a try-finally fashion.
This information is then used when handling errors. Error handling
will log this information and throw IOExcpetions.