First Running Version of Experiment Evaluator!
The first running version of the experiment evaluator has been created,
using a bogus/dummy module. An example class "EvaluationExamples"
combines the example data sets with the example document configurations
and applies different (so far only one) configurations of the Evaluator
to them.

This is a milestone in the project!