Small Set of Few but Diverse DocumentConfiguration Examples
In the class ExampleDocumentConfigurations, we now not just have the
potentially large list ALL_CONFIGURATIONS of many possible settings
and combinations of document drivers, graphics drivers, and chart
drivers. There now also is the list FEW_DIVERSE_CONFIGURATIONS, which
is suitable for examples requiring larger computational efforts. Both
lists are sorted approximately according to diversity, so that diverse
settings appear as early as possible.