Changed File Format for Color, Font, and Stroke Palettes to XML
The file (or resource) formats for color, font, and stroke palettes
have been changed to XML. Each format has been defined in a separate
XSD schema. All corresponding resources have been translated.

The reason for this change was that specifying font palettes in the
old format is tedious and hard to read. The new formats are clearer
and can be understood even without any documentation. Also they are
easily extensible, we can add more attributes or special types of
fonts, colors, or strokes relatively easily.