Added FileCollector and First Steps to LaTeX Document Driver
I included the class FileCollector as base class for
FileProducerSupport. Instances of FileProducerSupport are also instances
of java.io.Closeable and thus should not be handed out of the owning
scope. By making an abstract base class which offers the same functions
for adding files but which is not an instance of Closeable, we can
circumvent this problem.

A first few methods have been added to the LaTeX Document API backend.