"Almost-Working" Version of the LaTeX Implementation of the Document API
I now have a version of the LaTeX implementation of the Document API
which is almost working. Most of the basic functionality is there, but
there still are some bugs and some stuff is missing.

I also fixed a serious bug in the StreamEncoding class. Due to the lazy
loading of the default encodings, some code trying to detect the
encoding of a stream could create a default encoding before the real
instance of the default encoding was created. Later, when that instance
would actually be created, it would clash with the earlier one. Bang.
Anyway, that's solved now.