Fixed Citations and Underline for LaTeX Implementation of Document API
The citations which do not contain citation/number IDs have been fixed
to use the \nocite command in order to ensure that the reference is
included in the file.

I now use the ulem package to produce underline styles which can break
over lines. The \uline command is thus used instead of \underline. It
must be the inner-most command, i.e., we cannot do
  \uline{\texttt{...}}
but need to do
  \texttt{\uline{...}}