Fixed Tables for LaTeX Implementation of Document API
The tables created by the LaTeX implementation of the Document API now
look better, more like those from the XHTML 1.0 implementation of the
Document API.

However, multi-row cells will not work properly. They should be
rendered at the bottom with negative rowspan, but are rendered at
the top with positive rowspan. See
 http://texblog.org/2014/05/19/coloring-multi-row-tables-in-latex/
 http://tex.stackexchange.com/questions/200075/table-white-spacing-with-rowcolor-multicolumn-and-multirow
 
This issue has to be fixed in the future. I will ignore it for now.