Improved License and LaTeX Document API Implementation Package Loading
The package loading mechanism for the LaTeX Document API implementation
has been improved. We now automatically copy, e.g., the ulem.sty package
into the destination folder where the LaTeX document is created and
compiled. The goal of being able to automatically copy packages (from
project resources) is to ensure compatibility and to make sure that
the LaTeX code compiles even on systems where the package is not
available.

This, however, requires us to include some code as resources which may
be under a different license than the GPL used by our project. In order
to deal with this, the main license document has been modified to
explicitly list exceptions from the license terms.