IEEE Conference Document Class Added plus Minor BugFix
The IEEE conference document class (basically IEEEtran.cls with the
argument "conference") has been added. This document class looks the
almost same as as IEEEtran (as it is IEEEtran) but has a slightly
shorter text height (and maybe some other minor differences).

The EvaluationExamples class did not longer work due to an error in the
initialization order. This error was introduced as a side-effect of the
previous changes. It has been fixed.