LaTeX Implementation of Document API Now Supports Springer's llncs
The LaTeX implementation of the Document API now can generate LaTeX
documents with the document class llncs.cls of the Springer Verlag.
The document class and its required packages and BibTeX style are
included in the resources and automatically copied to the document
location. The LICENSE.md file has been updated to properly point out the
source of these resources and to state their copyrights (at least as far
as I could determine them).