Some Refactoring of the Chart API
The Chart API is now hosted as own utilities package instead of being
part of the graphics package. I think this makes more sense. Also some
of its code has been changed. The goal is to enable it to use smaller
fonts in the future, which is a bit of a hazzle.