Font Styles: Load Faces from Resources
The purpose of the FontStyle class is to offer a uniform interface
which allows us using the same fonts in graphics as well as in the
rest of the document. Ideally, documents generated by the Document API
should use uniform fonts and look approximately the same on all
computers.

We therefore already introduced the "face choices" list: The user can
specify a list of possible font faces to use for, e.g., "bold normal
font". The list is then processed one by one until a font is found which
is available on the current system. This font then is used. A similar
facility is provided by Cascading Style Sheets (CSS) in HTML.

This works more or less fine is we have full control over which fonts
are used. However, in case of the LaTeX implementation of the Document
API, we have not: LaTeX may have an own set of fonts which are not
accessible by Java. The fonts used in the document body text may thus
differ from those used in the graphics, which may look ugly.

I now added the functionality to provide fonts as resources in our jar.
The system can then first try to locate a font face on the system. If
the font face cannot be located, it can load the font from the resource
instead. This way we can increase the chance that document graphics and
document body use the same fonts.

This does not make the font face list facility obsolete, since there may
be some un-free fonts that we cannot provide (such as Helvetica). We
can only provide fonts under a free or open source license with our
code, such as the Nimbus Roman No9 L family.

Furthermore, it races the issue how to provide fonts with XHTML output.
I think there probably is a way to also reference font binaries from
CSS. There should be. I will check this later, and hopefully, if a font
is loaded from a resource, we can store this resource together with the
generated XHTML file (in case of the XHTML 1.0 Document API
implementation).