Workaround for a Very Odd Error in BibTeX Styles
There is a big problem with some BibTeX styles and characters with dot
accents (like "c" with a "." above). The correct way to get that would
be "{\.{c}}". However, some BibTex styles [1] eat away all dots,
leaving us with <code>{\{c}}</code>, which explodes. One bst style
triggering this error seems to be spbasic.bst, belonging to Springer's
svjour3.cls or llncs.cls.

We therefore load package "textcomp" providing command
"\captialdotaccent" (for capital letters) and add the following
definition to the LaTeX document's setup style:
"\gdef\lowerdotaccent#1{\.{#1}}%", which solves the problem for lower
case letters. This way, we can keep all "\." out of LaTeX produced for
BibTeX by replacing it with commands.

This is kind of a dirty hack, as it makes the bib files not portable
anymore. However, currently I don't have a better idea.

[1]
http://compgroups.net/comp.text.tex/problem-with-i-accent-in-bibtex/1921379