Two Bug Fixes: 1) Maven Pom and 2) Wrong Dependency
1) The Maven POM file was improved to have font files as non-filtered
resources. According to http://stackoverflow.com/questions/25404145/,
there otherwise may be problems: If font files get filtered, they may
be destroyed.

2) The example class RandomParallelExample for the Experiment Data I/O
had a dependency on org.junit.Assert which is only available during
testing but not in the compiled system. This only compiles under Eclipse
but fails during a Maven build (which I seemingly did not try for a
long time).