ReflectionUtils offers Search Path for Classes
ReflectionUtils now offers a search path for classes and instances.
Every time you call findClass, getInstanceByName, or
getResourceAsStream, you can specify a search path. The methods will
then first try to resolve the provided class names directly and if
that does not work, try to prepend the elements of the search path.

The goal is to make configuration easier. Let's say your program can
accept a parameter specifying a document driver from the Document API
to be used to generate the output. Usually, the path of a driver could
look like
 org.optimizationBenchmarking.utils.document.impl.xhtml10.XHTML10Driver
. This is quite long and inconvenient for a user to specify in a config
file. If we add
"org.optimizationBenchmarking.utils.document.impl."
to the search path, however, then it is enough to specify
"xhtml10.XHTML10Driver". If we even add
"org.optimizationBenchmarking.utils.document.impl.xhtml10.",
then the user just needs to specify "XHTML10Driver".