Font Sizes (and All Defaults) in Charts Now Managed by ChartDriver
All default values for fonts, strokes, and colors are now managed by
the ChartDriver, the abstract default implementation of the IChartDriver
interface of the Chart API. In particular, fonts are now automatically
resized to generate a visually pleasing experience. The reason is that
we want to use charts from within the Document API and this API uses
configurable FontPalettes. Now the font size of a font used in normal
text looks slightly too large when used in a chart. Thus, fonts used
in a chart should be resized in order to get a visually pleasing
experience. The ChartDriver implementation now does this. Since creating
a (resized) font may be an expensive operation, fonts are cached
internally by this class. Matter of fact, we cache all defaults for the
purpose of increasing speed.