BugFix for Issue with JFreeChart
We must subclass org.jfree.chart.plot.PiePlot3D to render our pie
charts because it ignores the method
org.jfree.chart.labels.PieSectionLabelGenerator#generateAttributedSectionLabel(PieDataset,Comparable).
This method is supposed to allow me to set the font for the labels.
However, it is never called. Thus, we need so tricks to get the fonts
into the labels.