Improved Support for XML File Types + Partial CSV Experiment Input
A major change has been done to XML support: The IXMLFileType interface
allows for specifying XML file types which provide a namespace URI and
a URL for downloading the schema.

A new parser for graphics drivers has been developed which can load a
default graphic driver for a given EGraphicFormat or load an instance
of a driver class.

IOJob now provides its Logger directly, which enables us to use it
elsewhere.

A partial and experimental CSV experiment data loader has been
implemented, but it does not work yet.