Modified Test Case for StableSum
The StableSumTest JUnit test sometimes fails in testSumUpLongs. This
seemingly happens extremely rarely, but is worrying me, as StableSum
should become one central part of the statistics module (to be
developed). I added more checks to this test case in the hope to track
down the problem. I also added test cases to check SaturatingAdd and
SaturatingSub.