Added Repository-Generation to pom.xml (and updated README.me a bit)
The pom.xml now can generate a Maven repository folder structure for the
project. This can then be copied to the gh-pages branch of the project.
This way, we can host the software and website and repository directly
in GitHub.

The readme file README.md has been updated a bit, too.