Begin of Work on More Unified Way to Load Experiment Data
We begin the work on a flattened wrapper around the hierarchical
experiment data creation API. This wrapper will make it more easy to
combine different file formats (such as EDI and CSV-based log files).
Actually, it will make parsing easier in general, as it will reduce
the amount of objects we need to deal with to 1 (at the cost of that
one object having more methods).