EDI Input Driver now based on FlatExperimentSetContext
The EDI input driver is now based on the FlatExperimentSetContext. We
can now extend it to allow mixing of EDI and other formats more easily.
The goal is to have a format which allows us to specify instances,
dimensions, as well as experiment properties in separate EDI files,
while storing the benchmark data (i.e., the data collected from the
single runs) in simple CSV files. This way, it becomes much easier to
use our suite and to create and edit the necessary data files.