Make Badges in README.me the Same Height (20px)
I made all the badges in the README.me the same height: 20px. The
Semaphore CI badge is originally only 18px high, which looks odd next to
all the 20px badges. However, now the font size in the Semaphore CI
badge is a bit bigger than for the other badges. Anyway, I think now it
looks a bit better.