EDI Input Now More Modular
The EDI Input is now more modular. The goal is to make it possible to
combine several different formats with EDI. For example, we want to be
able to store the log data as simple CSV files, while keeping the
definition of the experiment parameters, the dimensions, and the problem
instances in separate EDI files.

For this purpose, the new class EDIInputToolBase was created, which
provides the functionality for loading EDI files. Hopefully, it can be
extended to include, e.g., CSV capabilities.

The functionality of the EDIInput class stays basically the same, it is
a pure EDI loader.