CSV+EDI Experiment Set Data Input Now Working
The new CSVEIInput driver now works. This enables us to load data which
has experiment descriptions specified in EDI while having log files
which are simple CSV (or, better, space-separated-value) data lists.

The reason why this feature may be quite useful is that normally,
experiment data is generated as simple CSV format. In order to use
our tool suite, one therefore simply needs to specify the measurement
dimensions, the benchmark instance features, and the experiment
parameters in additional files. These files need to be in our EDI XML
format, an easy-to-understand and easy-to-validate XML format. Thus,
the additional effort on the researcher running the experiments is
limited.