Major Redesign of HashObject + Minor Changes
The class HashObject is intended to provide a simple base class for
objects which cache their hash code. Some objects may have lots of
data fields and computing their hash code time and again may not be
a good idea, performance wise. Thus, HashObject stores the hash code
and only calculates its once. In its original form, HashObject allowed
the user to provide a hash value and combined it with the hash code of
the object's class. This behavior is now removed, it just uses the user-
defined hash value. If we combine it with the class hash code as well,
some inconsistencies may inevitably result.

However, since many classes inherit from HashObject, this change may be
a bit problematic. Let's see how it works out.