Introduced SimplifyingGraphicProxy
Vector graphics formats such as EPS or PDF do not store pixel data.
Instead, they store the raw geometric primitives that were painted,
such as lines or rectangles or even text. This has the advantage that
you can zoom in without loss of fidelity and print them at an arbitrary
quality. Often, they are also smaller than pixel graphics such as PNG
or JPEG.

However, if your graphical primitives are complex, maybe you have many
lines or lines of many points, vector graphics become bigger. This is
the case if you render diagrams, such as Line Charts with the Chart API.
Even worse, it seems that the lines in these diagrams are not always
painted as poly-lines, i.e., a line of maybe 1000 points, but rather as
1000 separate line segments. This way, the ending point of one line
segment and the starting point of the following segment are the same,
i.e., the coordinates are stored twice, uselessly. This may potentially
waste quite some space (if you have many diagrams in a documents or just
several diagrams with many line segments).

I therefore tried to modify the graphics output API of my project to try
to reduce lines. The new class, SimplifyingGraphicProxy, tries to first
cache any "outline" objects you paint (such as lines, rectangles, poly-
lines, etc). Whenever necessary, e.g., if you want to change the paint
color, this cache is flushed. During this process, the class tries to
find if it can combine any line segments to a poly-line. If so, they are
rendered as poly-line, i.e., we can reduce the number of points to be
stored. It also detects redundant segments (i.e., if one segment appears
twice, and tries to omit them from the output.

This, of course, causes quite some overhead, both in terms of memory as
well as speed. It also makes only sense for vector graphics. Pixel (aka
Raster) graphics do not store geometric primitives, but only the final,
rendered pixel data. Such reduction will then not have any impact of the
image size.

The current version of this new concept is used as base class for the
vector formats I support. It is somewhat working, but still has some
issues to fix.