Fix of SimlifyingGraphicProxy for FreeHEP Graphic Drivers
The new version of the Graphic API tries to convert double or float
coordinates to integer coordinates, if this can be done without loss
of precision. A line from (70.0, 30.0) to (60.0, 20.0) would be plotted
as line from (70, 30) to (70, 20). I think that this is a good feature,
as it reduces the size of the data stored.

However, the FreeHEP implementations of vector graphics treat integer
coordinates sometimes a bit different from double coordinates: They
sometimes add a bias of 0.5 (see their class AbstractVectorGraphic).
My auto-conversion of coordinates will thus lead to some lines being
slightly off, sometimes.

Thus, I added the method autoConvertCoordinatesToInt(), which can
return false, in which case the coordinates are not converted.