Trying to Reduce File Size of FreeHEP Vector Graphics
I try to reduce the file sizes of the vector graphics produced by the
FreeHEP drivers by overriding "writeComment" to do nothing. This does
not bring much of a benefit: The SVG graphic class ignores the methods
and writes comments directly as strings to the output stream.
PSGraphics2D cannot be overridden as it basically loads a resource from
this.getClass() which does not exist when overridden, etc.

However, at least we can now properly log warnings to the Logger of the
graphics job.