Refactoring of Experiment Data API + Steps towards Property Grouping
Some minor refactoring of the Experiment Data API. For example, the
method "parameters()" of Experiment has been renamed to
"getParameterSetting()", which is more correct.

We also did some minor steps towards automatic grouping of Property
values. The goal is to be able to automatically group experiments
or instances according to their property values in a reasonable,
human-understandable way. This will make understanding evaluation
results much easier. But this code is far from working now.