Auto-Property-Value Grouping Now Fully Configurable
It is now possible to automatically group Instance Feature or
Experiment Parameter values, either by scale in terms of powers, e.g.,
[0, 1), [1, 10), [10, 100), [100, 1000) ... or by multiples of a
certain range, e.g., [0, 10), [10, 20), [20, 30), [30, 40) ...

This feature is fully configurable, i.e., one could go an define a
grouping for a given feature or parameter, such as stating that the
values of the feature "scale" should be grouped by powers of 2. Or
one could state that the values of a parameter populationSize should be
grouped in a way which will result in approximately between 3 and 7
groups.

The goal is to allow the user to customize the way we evaluate
experiment results.

I think that this is a pretty nice step forward.

Yet, this component is not yet realized in a "nice" way. The code
quality does not satisfy me yet, I think it will be a pain to maintain
or extend. For instance, we can only group by either distinct values
or numerical ranges. We cannot auto-group string parameters or anything
else. But well, for now it works (or at least seems to work) ^_^