PropertyValueGroup and PropertyValueGroups are now DataElements
Both, instances of PropertyValueGroup as well as their owners, the
instances of PropertyValueGroups, are now instances of DataElement.
This means they can be treated like elements of the Experiment Data
API and have own attributes computed based on them. This will make it
easier to compute (and cache) value-group specific statistics.