Made FreeHEP EPS Graphic Not Store Comments and Use Smaller Header
I found a work-around to prevent the FreeHEP-based EPS graphics to
store useless comments in the EPS files. I subclassed it and added a
header resource to the new subclass. This resource is very similar to
the original header, but is trimmed-down in size by removing comments
and other useless stuff, such as useless spaces and things. This will
make the generated EPS graphics (slightly) smaller.