Fix for Unwanted Log Output During JUnit Tests & Debug Configurations
I made some changes to the code to prevent the unwanted log output
during JUnit tests. I furthermore allowed for "Debug Configurations"
which are not linked to the root configuration. This makes sense for
JUnit testing, where we do not want to pollute the root context with
data from the each test case.