Added a Parser for Figure Sizes and Various Small BugFixes
I fixed some bugs, such as that the configuration actually did not hand
requests through the hierarchy when it could not find a value. Now it
does, so we can fully make use of the hierarchical way configurations
can be chained.

I added a parser for EFigureSize, which now allows us to conveniently
specify figure sizes in the configurations.

I furthermore fixed some odd issue in the evaluation configuration
process which led to an exception where none should occur (now it does
not anymore).

I think I fixed an issue with a forgotten closing tag in the
documentation of PropertyValueGrouperParser, which hopefully allows the
CodeShip builds to pass.

Finally, I added some default serialization (readResolve and
writeReplace) to some of the existing parsers.