Some Modifications to Basic Math Functions (to increase Accuracy)
I made some modifications to the basic math functions of the framework
which hopefully will increase the accuracy. The computations will be
slower. For example, when performing a double-precision addition,
subtraction, or multiplication, we will first check if we can cast all
arguments to long and whether the operation with the longs would be
possible without overflow. If so, we do it, because the results will
be 100% accurate. They are cast back to double, where then a loss of
accuracy may take place - but the resulting double would be the correct
double representation of the correct result. Only if no integer
arithmetic can be used, we resort to floating point arithmetic.