Tighter Coupling with Apache Commons Math3's FastMath
The FastMath class in Apache Commons Math3 states that it is a
  "Faster, more accurate, portable alternative to {@link Math} and
   {@link StrictMath} for large scale computation."

We therefore now rely more heavily on its methods in our mathematics
function classes. This dependency is made somewhat "soft", as we first
check whether this class is in the classpath and only use it if it is,
while otherwise using the default, java.lang.Math-based calculations.

In the future, this approach will probably be extended by using more
integer-based calculations (as in the basic arithmetic classes we
already have), and only resorting to double arithmetic if necessary.

If this can be done, I think we got the best and most accurate
arithmetics we can get under Java. ^_^