BugFix for Pow Function
The integer arithmetic of the implementation of the power function (Pow)
has been fixed. It should now better deal with overflows.

We now also try to use it even in the double precision version of the
pow function, if both arguments are actually long integer values and no
overflow would occur. This _should_ provide some additional accuracy
when the computation would exceed the mantissa of a double but still fit
within the 64 bits of a long. Maybe. Maybe not...