BugFix for FastMath.pow(double,int) Sometimes Returning NaN
It seems that FastMath.pow(double,int) sometimes returns NaN if the
second argument is large. We now check its results for this case and
use FastMath.pow(double,double) if a non-finite value is returned (such
as positive infinity or negative infinity or nan. This should do the
trick.