Is Integer Arithmetic Accurate + ColumnTransformedMatrix
I added the function "isLongArithmeticAccurate" to the class
MathematicalFunction. A function can now specify whether its long
arithmetic will return accurate results. For example, Min, Max,
and SaturatingAdd will always return accurate results when fed with
long arguments. Add will not, since the long arithmetic may overflow.
Log and Sin won't anyway, because they are not integer functions in the
first place. We can use this information to selectively choose whether
to calculate a function with long arithmetic or double arithmetic.

ColumnTransformedMatrix is a matrix class which can be put on top of
another matrix and applies a UnaryFunction to each column. This will
enable us to translate a list of coordinate points, say X, Y to log X,
Y on the fly. This new class makes use of the above 
"isLongArithmeticAccurate" function. It's getLong and getDouble methods
use it to find whether to calculate its result via computeAsLong or
computeAsDouble.