First Step of Experiment Data API Abstraction
The Experiment Data API is designed for high performance data access and
highest possible data integrity and fidelity. In order to achieve
lossless data storage with high access speed when dealing with both
long and double data at the same time in the same data points, it will,
e.g., on-the-fly compile suitable data structures. It also is designed
for fast access of properties and a high data integrity, achieved by
internal generation of unique IDs and orderings.

This will benefit the way in which we evaluate data, however, it also
forces me to make the API rather static. You cannot easily generate a
new instance, of, say, "Experiment" without violating some constraint.
However, especially when computing more complex statistics, it is
desirable to be able to dynamically group parts of the data. We can
already do that via PropertyValueGrouper. However, the result is a new
data structure, incompatible with the structures of the Experiment Data
API. Thus, if you want to compute a statistic over a set of runs, you
would have to implement it multiple times, for different cases. This is
not good.

In order to be able to both benefit from rigid data integrity
constraints while being able to generate new instances of the Experiment
Data API on the fly, I have begun to create an abstract "specification"
of data structures in form of interfaces. These interfaces are then
implemented by the "reference implementation", which is the previous
version of the Experiment Data API.

Using interfaces will allow me to easily introduce new kinds of data
elements whenever needed. The speed and fidelity of the original API is
preserved (except that we will now do interface calls instead of final
method calls, which is slower).

Another benefit from this abstraction is that it will make it easier
for new users to understand the data structures we use.

Still, all implementations of the interfaces (except for IDataPoint)
must trace back to one common base class, "DataElement". This is
necessary, because otherwise I cannot realize dynamic computation and
caching of attributes, which is very desirable to speed up big
computations.