Evaluator Now Uses New Interfaces of Experiment Data API
The specification and implementation of the Evaluator component now use
the new interfaces of the Experiment Data API. I continue in my step by
step transition from using the reference implementation objects of the
API to just using the abstract API. This will make understanding and
maintaining the code easier. But it will probably lead to some dead code
and other ugly stuff lying around for some time.