The JUnit Test ExperimentSetTest Now Also Uses The New Interfaces
The JUnit test ExperimentSetTest, the test with which I try to check
whether the implementation of the Experiment Data API does what it is
supposed to do, is now also based on the new interfaces.

I just abstracted and generalized the Experiment Set API by specifying
its functionality in a set of interfaces. The original objects of the
API now implement these interfaces. On the long run, this allow us to
have cleaner and more elegant code. Foremost, we can keep the many
internal classes of the Experiment Data API in the reference
implementation package without worrying too much about them. All our
coding can concentrate on the interfaces, which are clean and nice. The
actual implementation can take care of robustness, fidelity, and
performance behind the scenes.