ArrayListView instead of ArraySetView in Experiment Data API
The Experiment Data API now returns instances of ArrayListView as the
default data set collection representation instead of ArraySetView. The
reference data implementation uses and has used ArraySetView, which is
potentially much faster when looking up elements. However, it requires
its elements to be able to meaningful comparison. This is possible in
the reference implementation because of its rigid data integrity and
indexing mechanisms. When implementing delegates or subsets of data
elements, it is, however, hard to achieve this with reasonable effort.
Thus, the reference implementation can use the faster data structures,
which other implementations of the Experiment Data API may just use
normal lists to represent data.