Redesign of Delegating Experiment Data API Elements Necessary
A re-design of the delegating elements of the Experiment Data API will
be necessary. Originally, the idea was that we can choose, e.g., a sub-
set of instances and calculate statistics over them. By doing so, as
many attribute computations as possible should be forwarded to the
original objects. This would greatly increase speed and reduce memory
requirements for caching.

However, if we provide this sub-set with the full capabilities of
IExperimentSet, this may lead to inconsistencies. If somebody tries to
access the IInstance of an instance runs sub-set and obtains an
attribute for this object, the attribute would be computed either on
the full instance runs sets (and thus be wrong) or computed on the sub-
sets alone and cached in the original instance, which would be wrong,
too, depending on the implementation.

I will either need to reduce the capabilities of the sub-sets or need
to introduce a more comprehensive "shadowing" approach.