More Shadows of Experiment Data API
A basic set of shadow objects for the Experiment Data API now exist.
A "shadow" of an element of the API is an object which shares some
properties with the original but may be a "subset" of that object. For
example, the shadow of an IExperiment could be an IExperiment with only
a subset of the instance runs of the original experiment, but with the
same parameters and name. Creating such subsets may be a useful thing
when grouping data together or dividing data elements.

The problem is that implementing the interface IExperiment entails
implementing the whole hierarchy of IExperimentSet, with IFeature,
IInstance, and all of that, because these elements form a
bi-directional graphic based on getOwner() and getData(). This
structure is necessary for having convenient data access, however it
would mean to fully copy the complete original data whenever wanting to
make a subset as described above.

The shadow objects, however, may omit copying some data initially and
creating shadows of that data when needed. If you make a subset of
an original experiment, you need to shadow the IExperiment, the
IExperimentSet, and the elements of the subset. Shadows of the other
objects, such as IInstance or IFeature or IParameter will be created
when needed. This should safe runtime and memory, while preserving the
original interface semantics.