Experiment Set Shadowing Now Fully JUnit Tested and Passing
The shadows of the whole experiment sets can now pass the same JUnit
tests as the reference implementation objects they are shadowing. This
means they should be fully compatible (at least I hope that). If this
is true, then they can be used to build new subsets and groups or
even supersets of existing data.

The attribute computation and caching is currently only shared for
IRuns and partly shared for IInstanceRuns and IExperiment. A more
efficient "sharing" of values computed on shadow objects and their
originals would be nice. But for now, things will do as they are.