First Concept of an ECDF Attribute
The first version or concept of an attribute for computing the
Estimated Cumulative Distribution Function (ECDF) over an
IExperimentSet, IExperiment, IInstanceRuns, or IRun has been introduced.
The value of the attribute will be a floating point matrix
(DoubleMatrix1D), since the fraction of succeeded runs is, well, a
fraction, and hence a floating point number anyway. During the
computation, however, we use the most precise data representation for
both the time and goal dimension.