New Interfaces for Clustering and Improvement to Aggregates
Some new interfaces for clustering have been introduced in order to
generalize the way experiments and instance runs can be grouped. This
will allow us to later use different clustering techniques. Right now
we only support grouping based on feature and parameter values.

PathUtils has been extended with a method to sanitize path components in
order to ensure that the path suggestions in the Document API are always
valid. This new method is used in the abstract implementation base
classes of the document API.

In some of the statistics and aggregate classes, we now use the methods
of NumericalType more heavily to detect whether to use long or double
arithmetic.